main/requirements.txt
# ============================================================
#  NET-1 — Python Dependencies (Production)
# ============================================================

# Static files (production)
whitenoise
asgiref==3.11.1
sqlparse==0.5.5
gunicorn

# Database
psycopg2-binary==2.9.11

# Cache / Valkey
django-valkey==0.4.0
django-redis==6.0.0
valkey==6.1.1
redis==7.3.0

# LLM
ollama==0.6.1

# HTTP
requests==2.32.5
httpx==0.28.1
httpcore==1.0.9
h11==0.16.0
certifi==2026.2.25
charset-normalizer==3.4.5
idna==3.11
urllib3==2.6.3

# PDF thumbnails
pdf2image==1.17.0
pillow==12.1.1

# System monitoring
psutil==7.2.2

# Environment / config
python-decouple==3.8

# Utilities
python-dateutil==2.9.0.post0
six==1.17.0
tzdata==2025.3
tzlocal==5.3.1
packaging==26.0
typing_extensions==4.15.0
pydantic==2.12.5
pydantic_core==2.41.5
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
prompt_toolkit==3.0.52
wcwidth==0.6.0
